refactor: parse_scenario function and add more testing
The parse_scenario function in assistant.rs has been refactored to enhance its robustness. It now checks if pairs of prompts and workflow match before returning results. In addition to that, new structs Request and Response have been created to be used as Renderer parameters, instead of the previous empty struct R. Furthermore, a significant number of tests were added to verify the new implementation of the parse_scenario function. This increases the reliability of the function and makes the code more maintainable.